Rename containers components and styles files
